feat(router): route deis.domain to controller
Forwards requests to deis.domain to the controller.

TESTING: rebuild the router and controller with this change:

    $ make -C router build restart
    $ make -C controller build restart

Test that deis.domain routes to the controller and that other
requests to the domain route to the app. Also, send several requests
to the API and be sure that nothing broke related to auth because
of the proxying of requests.

The integration tests were updated to test the controller and the app
routing, so you can also `cd test && bundle exec rake`.

Also, try creating an app with the name of 'deis' - it should be
rejected.

replaces #974
part of #535