feat(version): add version package
This package is a central version number for all Deis components
written in Go. If a package needs to refer to the current version
of Deis, they can follow this convention:

```
package main

import (
    "fmt"

    "github.com/deis/deis/version"
)

func main() {
    fmt.Printf(version.Version)
}
```

This will be especially handy for handling more than one component
relying on the same version number e.g. the `deisctl` and the
`deis` client (if it's rewritten in Go, which is assumed to be on
the roadmap).